refactor(api): change coordinate types from float32 to float64 for precision

+++ website/api/optimiser/_models/models.go
@@ -131,8 +131,8 @@ func (slot *ModuleSlot) ParseModuleSlotFields(lessonKey string) error {
 }
 
 type Coordinates struct {
-	X float32 `json:"x"`
-	Y float32 `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type Location struct {
